Ensures that uprof can be used to profile dynamicly loaded modules.
Modules that can be unloaded pose some problems since we can't rely on the
persistance of static global variables and various constants such as
__FILE__ or __FUNCTION__.

We still use static globals for timers and counters but we now hang a heap
allocated ->state structure within them. When modules are first loaded this
is NULL, so whenever we see a timer with a NULL state we search for a
corresponding state structure using the timers name, and if not found we
allocate a new one.

All statistics are accumulated in the state structure, so nothing is lost
if the module unloads, and if the module is reloaded we will resume
updating the original state structures.