Adds a way to relate ("link") stats from multiple contexts
If you are adding profiling to a dynamically loadable module that may be
loaded in multiple contexts and you can't depend on a shared context symbol
then you can instead choose to link contexts at runtime.

For example this has been used to add DRI driver instrumentation allowing
timers to have parent names that are only relevent when loaded for direct
rendering via libGL as opposed to the xserver.