Adds a way to queue messages that will be output at report time
The idea is to be able to group together a list of advice that may give
insight into performance issues affecting an application. E.g. for
Clutter we might report things like:

"You appear to be slicing a huge amount of textures!", or
"You are downloading textures from the GPU very frequently!", or
"Cogl isn't able to batch your geometry very well"