Fix building on the x86-64 architecture
In x86-64 unsigned long long is not considered the same type as
uint64_t so it was complaining that the declaration for
uprof_get_system_counter did not match the definition. It was also
throwing warnings where %llu was used to format a 64-bit value.

As glib is used throughout the program, it might as well use guint64
to get a 64-bit type and then we can use G_GUINT64_FORMAT to get the
format specifier for it.