Improve the context suspend/resume semantics
A context can now be suspended multiple times and will need an equal number
of resumes to fully resume.

When suspending a timer that is currently timing we will keep track of the
duration it has been timing so far and continue to add to this duration if
the timer is resumed before it is stopped.  If it is stopped before being
resumed then we will still increment the counter of the timer and record the
duration up to the point of becoming suspended.

Any timer that is started while the context is suspended will get started
automatically when the context is resumed.