implements uprof_report_add_{counters,timers}_attribute
This allows developers to register custom attributes for their counter
and timers which will add extra columns to the printed report. A given
callback will be called for each counter/timer result and it should
return the attribute value as a string.

The values for custom attributes should print and align correctly even
if the value string is wrapped onto multiple lines or contains
multi-byte utf8 characters.