development release: Bump the version to 0.3
The new dbus interface the uprof_report_add_{counters,timers}_attribute
functions and the ncurses UI for displaying live statistics warrant a
new release for projects to depend on. I'd like to get a bit of feedback
and time to do some bug fixing before releasing 0.4 so this is an
initial development release.