dbus: This adds DBus support to UProf
This makes uprof create a dbus object for each UProfReport currently
instantiated that tools can then request report data from.

When uprof is initialized a /org/freedesktop/UProf/Service object is
registered with the application's session bus connection. This object
implements the org.freedesktop.UProf.Service interface and provides a
means to discover the names of other objects that implement the
org.freedesktop.UProf.Reportable interface.

When a UProfReport is created via uprof_report_new ("Example Report")
then a dbus object like /org/freedesktop/UProf/Reports/Example_Report is
created implementing the .UProf.Reportable interface.

Currently the only method available for the UProf.Service interface it
"ListReports" which returns a string array of report names. The only
method available for the UProf.Reportable interface is GetTextReport
which returns a string with a fully formatted report that can be printed
to a terminal.

A simple "uprof-tool" commandline tool has been added that can be used
as a client to fetch text reports from .UProf.Reportable objects.