report: Add support for custom statistics
This allows the addition of custom statistics to a report. Each
statistic has a name and list of associated attributes (which
corresponds to columns in a text report) A callback can be given to
define the value of an attribute.

Note: currently the statistics can only be associated with a report
not a context so if an application were to link in the context of
one of its dependency libraries into its own context then any
report of that context wont contain any custom statistics.