report: re-adding a custom statistic should silently ignore
Asserting that the statistic name doesn't already exist makes it more
awkward for code using the uprof_report_add_statistic API since it would
have to keep track of the first time the statistic is added to be sure
it wont hit an assert later. Now we simply update the description if a
statistic name is re-added.