Adds ui and further support for trace messages
This adds a second page to the uprof-tool ui that shows traces signals
from the current reportable object.

This makes a tweak to the message format so now they should look like:

"[category1,category0,categoryA] filename:line:function & message"

This will allow potential extensions to the list of ':' separated
fields.